pipeline/requirements.txt
neuprint-python>=0.5.0
navis>=1.6.0
pandas
numpy
trimesh
